refactor(lns-cli): move config args (ConfigArgs/ConfigCommand/…) into the config module
